backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
supabase==1.0.3
python-multipart==0.0.6
httpx==0.25.1
numpy==1.26.0
transformers==4.34.1
pyaudio==0.2.13
python-jose==3.3.0
passlib==1.7.4
